Check point work on disabling tags

+++ test/qcbor_decode_tests.c
@@ -6070,7 +6070,7 @@ int32_t CBORTestIssue134()
 
    uCBORError = QCBORDecode_Finish(&DCtx);
 
-   return uCBORError;
+   return (int32_t)uCBORError;
 }
 
 int32_t CBORSequenceDecodeTests(void)
